requirements.txt
alembic==1.4.3
aniso8601==8.0.0
apispec==4.0.0
APScheduler==3.6.3
astroid==2.4.2
attrs==20.1.0
awscli==1.18.205
bcrypt==3.2.0
blinker==1.4
boto==2.49.0
boto3==1.16.45
botocore==1.19.45
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
docutils==0.15.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.2
phonenumbers==8.12.11
Pillow==8.0.0
psycopg2==2.8.5
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.5
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
twilio==6.45.1
tzlocal==2.1
urllib3==1.25.10
webargs==7.0.1
Werkzeug==1.0.1
wrapt==1.12.1
